Make branch distance work
When working on Advent of Code, the distances were measured in terms
of orbits; specifically, the distance wasn't from the endpoints, but
the nodes the endpoints were attached to.

So far as I can tell, that's not the standard way to measure distances
in trees, so update find/{2,3} to include the full path to the node,
including the node's label itself, and add/update tests to suit.